split extension into files
